requirements.txt
gradio>=4.0.0
openai>=1.0.0
python-dotenv>=1.0.0
pytest>=8.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
prometheus_fastapi_instrumentator>=6.1.0
fastapifastapi>=0.110.0,<1.0.0
pydantic>=2.6,<3.0
uvicorn[standard]>=0.28,<1.0